refactor: rename http_mock_config_capture_test to http_mocker_config_capture_test and extract API spec initialization
