+++ internal/testutil/unit/http_mocker_config_capture_test.go
@@ -14,7 +14,6 @@ import (
 )
 
 var (
-
 	//go:embed testdata/CaptureTest/createCluster.json
 	createClusterReqBody string
 	//go:embed testdata/CaptureTest/createClusterResponse.json
@@ -32,6 +31,7 @@ func TestFailedFilename(t *testing.T) {
 
 func TestCaptureMockConfigClientModifier_clusterExample(t *testing.T) {
 	t.Setenv(unit.EnvNameHTTPMockerCapture, "true")
+	unit.InitializeAPISpecPaths()
 	clientModifier := unit.NewCaptureMockConfigClientModifier(t, 2)
 	transport := httpmock.NewMockTransport()
 	client := http.Client{Transport: transport}
